working on Rpi time stamps

+++ apps/senspot/senspot.c
@@ -37,7 +37,7 @@ void SenspotPrint(SENSPOT *spt, unsigned long seq_no)
 
 	tm.tv_sec = ntohl(spt->tm.tv_sec);
 	tm.tv_usec = ntohl(spt->tm.tv_usec);
-	ts = (double)((unsigned long)tm.tv_sec) + ((double)((unsigned long)tm.tv_usec) / 1000000.0);
+	ts = (double)((unsigned int)tm.tv_sec) + ((double)((unsigned int)tm.tv_usec) / 1000000.0);
 
 	fprintf(stdout,"time: %10.10f ",ts);
 	fprintf(stdout,"%s ",spt->ip_addr);
